made buttons npt crash program

+++ GUIApplication/GUIApplication/MainWindow.xaml.cs
@@ -43,14 +43,20 @@ namespace GUIApplication
             if (buyerTab.IsSelected)
             {
                 Buyer buyer = (Buyer)buyerData.SelectedItem;
-                Location loc = iServ.GetLocation(buyer.ZipCode);
-                MessageBox.Show("KøberID: " + buyer.Id + "\nNavn: " + buyer.Name + "\nAdresse: " + buyer.Address + "\nPostnummer: " + buyer.ZipCode + " By: " + loc.City + "\nTelefon: " + buyer.Phone + "\nMobil: " + buyer.Mobile + "\nEmail: " + buyer.Email + "\nMisc: " + buyer.Misc);
+                if (buyer != null)
+                {
+                    Location loc = iServ.GetLocation(buyer.ZipCode);
+                    MessageBox.Show("KøberID: " + buyer.Id + "\nNavn: " + buyer.Name + "\nAdresse: " + buyer.Address + "\nPostnummer: " + buyer.ZipCode + " By: " + loc.City + "\nTelefon: " + buyer.Phone + "\nMobil: " + buyer.Mobile + "\nEmail: " + buyer.Email + "\nMisc: " + buyer.Misc);
+                }
             }
             else
             {
                 Seller seller = (Seller)sellerData.SelectedItem;
-                Location loc = iServ.GetLocation(seller.ZipCode);
-                MessageBox.Show("SælgerID: " + seller.Id + "\nNavn: " + seller.Name + "\nAdresse: " + seller.Address + "\nPostnummer: " + seller.ZipCode + " By: " + loc.City + "\nTelefon: " + seller.Phone + "\nMobil: " + seller.Mobile + "\nEmail: " + seller.Email + "\nMisc: " + seller.Misc);
+                if (seller != null)
+                {
+                    Location loc = iServ.GetLocation(seller.ZipCode);
+                    MessageBox.Show("SælgerID: " + seller.Id + "\nNavn: " + seller.Name + "\nAdresse: " + seller.Address + "\nPostnummer: " + seller.ZipCode + " By: " + loc.City + "\nTelefon: " + seller.Phone + "\nMobil: " + seller.Mobile + "\nEmail: " + seller.Email + "\nMisc: " + seller.Misc);
+                }
             }
         }
 
@@ -84,18 +90,24 @@ namespace GUIApplication
             if (buyerTab.IsSelected)
             {
                 Buyer buyer = (Buyer)buyerData.SelectedItem;
-                UpdateBuyer window = new UpdateBuyer(buyer);
-                window.WindowStartupLocation = WindowStartupLocation.CenterScreen;
-                window.Topmost = true;
-                window.Show();
+                if (buyer != null)
+                {
+                    UpdateBuyer window = new UpdateBuyer(buyer);
+                    window.WindowStartupLocation = WindowStartupLocation.CenterScreen;
+                    window.Topmost = true;
+                    window.Show();
+                }
             }
             else
             {
                 Seller seller = (Seller)sellerData.SelectedItem;
-                UpdateSeller window = new UpdateSeller(seller);
-                window.WindowStartupLocation = WindowStartupLocation.CenterScreen;
-                window.Topmost = false;
-                window.Show();
+                if (seller != null)
+                {
+                    UpdateSeller window = new UpdateSeller(seller);
+                    window.WindowStartupLocation = WindowStartupLocation.CenterScreen;
+                    window.Topmost = false;
+                    window.Show();
+                }
             }
         }
 
@@ -103,12 +115,15 @@ namespace GUIApplication
         {
             if (buyerTab.IsSelected)
             {
-                 MessageBoxResult messageBoxResult = System.Windows.MessageBox.Show("Er du sikker?", "Slet køber", System.Windows.MessageBoxButton.YesNo);
+                MessageBoxResult messageBoxResult = System.Windows.MessageBox.Show("Er du sikker?", "Slet køber", System.Windows.MessageBoxButton.YesNo);
                 if (messageBoxResult == MessageBoxResult.Yes)
                 {
                     Buyer buyer = (Buyer)buyerData.SelectedItem;
-                    iServ.DeleteBuyer(buyer);
-                    buyerData.ItemsSource = iServ.GetAllBuyers();
+                    if (buyer != null)
+                    {
+                        iServ.DeleteBuyer(buyer);
+                        buyerData.ItemsSource = iServ.GetAllBuyers();
+                    }
                 }
             }
             else
@@ -117,8 +132,11 @@ namespace GUIApplication
                 if (messageBoxResult == MessageBoxResult.Yes)
                 {
                     Seller seller = (Seller)sellerData.SelectedItem;
-                    iServ.DeleteSeller(seller);
-                    sellerData.ItemsSource = iServ.GetAllSellers();
+                    if (seller != null)
+                    {
+                        iServ.DeleteSeller(seller);
+                        sellerData.ItemsSource = iServ.GetAllSellers();
+                    }
                 }
             }
         }
@@ -188,14 +206,14 @@ namespace GUIApplication
             Appointment ap = (Appointment)appointmentData.SelectedItem;
             string formatDate = "dd.MM-yy";
             string formatTime = "HH.mm";
-            string info = "ID: " + ap.Id + "\nDato: " + ap.Date.ToString(formatDate) + "\tVarighed: " 
+            string info = "ID: " + ap.Id + "\nDato: " + ap.Date.ToString(formatDate) + "\tVarighed: "
                    + ap.StarTime.ToString(formatTime) + " - " + ap.EndTime.ToString(formatTime) + "\nKategori: "
                    + ap.Category + "\nBeskrivelse: " + ap.Description + "\nStatus: " + ap.Status;
-            if(ap.Buyer != null)
+            if (ap.Buyer != null)
             {
                 info += "\nKunde(køber): " + ap.Buyer.Name;
             }
-            else if(ap.Seller != null)
+            else if (ap.Seller != null)
             {
                 info += "\nKunde(sælger): " + ap.Seller.Name;
             }
